+++ src/app/signup/page.tsx
@@ -1,10 +1,10 @@
-import AuthForm from "@/components/Form";
+import AuthForm from "@/components/(login,signup)/Form";
 import React from "react";
 
 const SignUpPage = () => {
   return (
     <div className="flex flex-row justify-center items-center bg-black ">
-      <AuthForm mode="signup" />;
+      <AuthForm mode="signup" />
     </div>
   );
 };
